requirements.txt
Flask==3.0.3
flask_wtf
WTForms==3.1.2
python_dotenv==1.0.1